pyproject.toml
[project]
name = "kambacode-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-compressor>=4.5.1",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-ninja>=1.4.5",
    "django-ninja-jwt[crypto]>=5.4.0",
    "django-vite>=3.1.0",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.11.0",
]
